+++ locales/en.json
@@ -1,0 +1,45 @@
+{
+  "app_title": "Git2Base",
+  "file_menu": "File",
+  "edit_menu": "Edit",
+  "help_menu": "Help",
+  "repo_selection": "Repository Selection",
+  "select": "Select",
+  "branch_selection": "Branch Selection",
+  "refresh_branches": "Refresh Branches",
+  "commit_density": "Commit Density",
+  "show_current_branch_only": "Show Current Branch Only",
+  "refresh_commits": "Refresh Commits",
+  "commit_list": "Commit List",
+  "commit1": "Commit 1",
+  "commit2": "Commit 2",
+  "get_all_commits": "Get All Commits",
+  "get_commits_after": "Get Commits After",
+  "compare_commits": "Compare Commits",
+  "compare_and_analyze": "Compare and Analyze",
+  "analyze_commits_diff": "Analyze Commits Diff",
+  "command_preview": "Command Preview",
+  "execute": "Execute",
+  "reset_db": "Reset Database",
+  "command_output": "Command Output",
+  "select_repo_title": "Select Git Repository Directory",
+  "error": "Error",
+  "invalid_repo": "Invalid Git Repository",
+  "warning": "Warning",
+  "select_valid_repo": "Please select a valid Git repository",
+  "select_branch": "Please select a branch",
+  "enter_commit_hash": "Please enter a commit hash",
+  "enter_two_commit_hashes": "Please enter two commit hashes",
+  "refresh_branches_failed": "Failed to refresh branches",
+  "get_commits_failed": "Failed to get commits",
+  "density_options.all": "All",
+  "density_options.1week": "1 Week",
+  "density_options.2weeks": "2 Weeks",
+  "density_options.1month": "1 Month",
+  "density_options.3months": "3 Months",
+  "density_options.6months": "6 Months",
+  "density_options.1year": "1 Year",
+  "confirm_reset_db": "Are you sure you want to reset the database? This will delete all analysis data!",
+  "no_command": "No command to execute",
+  "command_failed": "Command execution failed"
+}
