+++ locales/ja.json
@@ -1,0 +1,45 @@
+{
+  "app_title": "Git2Base",
+  "file_menu": "ファイル",
+  "edit_menu": "編集",
+  "help_menu": "ヘルプ",
+  "repo_selection": "リポジトリ選択",
+  "select": "選択",
+  "branch_selection": "ブランチ選択",
+  "refresh_branches": "ブランチを更新",
+  "commit_density": "コミット密度",
+  "show_current_branch_only": "現在のブランチのみ表示",
+  "refresh_commits": "コミットを更新",
+  "commit_list": "コミット一覧",
+  "commit1": "コミット1",
+  "commit2": "コミット2",
+  "get_all_commits": "すべてのコミットを取得",
+  "get_commits_after": "指定コミット以降を取得",
+  "compare_commits": "コミットを比較",
+  "compare_and_analyze": "比較と分析",
+  "analyze_commits_diff": "コミット差分を分析",
+  "command_preview": "コマンドプレビュー",
+  "execute": "実行",
+  "reset_db": "データベースをリセット",
+  "command_output": "コマンド出力",
+  "select_repo_title": "Gitリポジトリディレクトリを選択",
+  "error": "エラー",
+  "invalid_repo": "無効なGitリポジトリ",
+  "warning": "警告",
+  "select_valid_repo": "有効なGitリポジトリを選択してください",
+  "select_branch": "ブランチを選択してください",
+  "enter_commit_hash": "コミットハッシュを入力してください",
+  "enter_two_commit_hashes": "2つのコミットハッシュを入力してください",
+  "refresh_branches_failed": "ブランチの更新に失敗しました",
+  "get_commits_failed": "コミットの取得に失敗しました",
+  "density_options.all": "すべて",
+  "density_options.1week": "1週間",
+  "density_options.2weeks": "2週間",
+  "density_options.1month": "1ヶ月",
+  "density_options.3months": "3ヶ月",
+  "density_options.6months": "6ヶ月",
+  "density_options.1year": "1年",
+  "confirm_reset_db": "データベースをリセットしますか？この操作によりすべての分析データが削除されます！",
+  "no_command": "実行可能なコマンドがありません",
+  "command_failed": "コマンドの実行に失敗しました"
+}
