+++ locales/zh.json
@@ -1,0 +1,45 @@
+{
+  "app_title": "Git2Base",
+  "file_menu": "文件",
+  "edit_menu": "编辑", 
+  "help_menu": "帮助",
+  "repo_selection": "仓库选择",
+  "select": "选择",
+  "branch_selection": "分支选择",
+  "refresh_branches": "刷新分支",
+  "commit_density": "提交密度",
+  "show_current_branch_only": "仅显示当前分支",
+  "refresh_commits": "刷新提交",
+  "commit_list": "提交列表",
+  "commit1": "提交1",
+  "commit2": "提交2",
+  "get_all_commits": "获取所有提交",
+  "get_commits_after": "获取指定提交后的提交",
+  "compare_commits": "比较提交",
+  "compare_and_analyze": "比较并分析",
+  "analyze_commits_diff": "分析提交差异",
+  "command_preview": "命令预览",
+  "execute": "执行",
+  "reset_db": "重置数据库",
+  "command_output": "命令输出",
+  "select_repo_title": "选择Git仓库目录",
+  "error": "错误",
+  "invalid_repo": "无效的Git仓库",
+  "warning": "警告",
+  "select_valid_repo": "请选择有效的Git仓库",
+  "select_branch": "请选择分支",
+  "enter_commit_hash": "请输入提交哈希",
+  "enter_two_commit_hashes": "请输入两个提交哈希",
+  "refresh_branches_failed": "刷新分支失败",
+  "get_commits_failed": "获取提交失败",
+  "density_options.all": "全部",
+  "density_options.1week": "1周",
+  "density_options.2weeks": "2周",
+  "density_options.1month": "1个月",
+  "density_options.3months": "3个月",
+  "density_options.6months": "6个月",
+  "density_options.1year": "1年",
+  "confirm_reset_db": "确定要重置数据库吗？此操作将删除所有分析数据！",
+  "no_command": "没有可执行的命令",
+  "command_failed": "命令执行失败"
+}
